+++ src/main/scala/TwitterStreamer.scala
@@ -3,36 +3,22 @@
  * NOTE: this may stop working if at any point Twitter does some breaking changes to this API or the JSON structure.
  */
 
-import akka.NotUsed
 import akka.actor.ActorSystem
 import akka.http.scaladsl.Http
 import akka.http.scaladsl.model.HttpHeader.ParsingResult
 import akka.http.scaladsl.model._
-import akka.stream.{ActorMaterializer, ClosedShape}
-import akka.stream.scaladsl.{Broadcast, Flow, GraphDSL, RunnableGraph, Sink, Source}
+import akka.stream.ActorMaterializer
+import akka.stream.scaladsl.Flow
 import akka.util.ByteString
 import com.hunorkovacs.koauth.domain.KoauthRequest
 import com.hunorkovacs.koauth.service.consumer.DefaultConsumerService
 import org.json4s._
-import org.json4s.native.JsonMethods._
+import tweet.Stream
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
-import scala.concurrent.duration._
-import scala.util.{Failure, Success, Try}
 
 object TwitterStreamer extends App {
-
-  case class TweetCounts(original: Int, retweet: Int) {
-    val originality = if (original + retweet > 0) {
-      100 * original / (original + retweet)
-    } else {
-      0
-    }
-
-    def increment(count: TweetCounts): TweetCounts = TweetCounts(original + count.original, retweet + count.retweet)
-  }
-
 	//Get your credentials from https://apps.twitter.com and replace the values below
 	private val consumerKey = sys.env("CONSUMER_KEY")
 	private val consumerSecret = sys.env("CONSUMER_SECRET")
@@ -40,9 +26,9 @@ object TwitterStreamer extends App {
 	private val accessTokenSecret = sys.env("ACCESS_TOKEN_SECRET")
 	private val url = "https://stream.twitter.com/1.1/statuses/filter.json"
 
-	implicit val system = ActorSystem("twitter-streamer")
-	implicit val materializer = ActorMaterializer()
-	implicit val formats = DefaultFormats
+  implicit val system = ActorSystem("twitter-streamer")
+  implicit val materializer = ActorMaterializer()
+  implicit val formats = DefaultFormats
 
 	private val consumer = new DefaultConsumerService(system.dispatcher)
 
@@ -96,57 +82,13 @@ object TwitterStreamer extends App {
           response.entity.dataBytes.runForeach{ x => println(x.utf8String)}
         } else {
           val source = response.entity.dataBytes.via(extractFromFireHose)
-          Future(mainStream(source))
+          Future(Stream.mainStream(source))
         }
       }
     }
   }
 
-  def jsonExtract(json: String): Try[Tweet] = Try(parse(json).extract[Tweet]) match {
-    case Success(t) => Success(t)
-    case Failure(e) =>
-      println("-----")
-      println(e.getMessage)
-      println(json)
-      Failure(e)
-  }
-
-  def extractTweet: Flow[String, Tweet, Any] = Flow[String]
-    .map(jsonExtract)
-    .collect { case Success(t) => t }
-
-  def saveTweet: Sink[Tweet, Any] = Sink.ignore
-
-  def tweetStats: Flow[Tweet, String, Any] = Flow[Tweet]
-    .map { tweet => if (tweet.retweeted_status.isDefined) { TweetCounts(0,1) } else { TweetCounts(1,0) } }
-    .groupedWithin(20, 10.seconds)
-    .map { tweetCounts =>
-      tweetCounts.reduceLeft(_.increment(_))
-    }
-    .statefulMapConcat{ () =>
-      var totalCount = TweetCounts(0,0)
-      count => {
-        totalCount = totalCount.increment(count)
-        List(totalCount)
-      }
-    }
-    .map { counts =>
-      s"${counts.original} original tweets and ${counts.retweet} retweets. Originality: ${counts.originality}%"
-    }
-
-  def mainStream(source: Source[String, Any]) = {
-    val g = RunnableGraph.fromGraph(GraphDSL.create() { implicit builder: GraphDSL.Builder[NotUsed] =>
-      import GraphDSL.Implicits._
-
-      val bcast = builder.add(Broadcast[Tweet](2))
-      source ~> extractTweet ~> bcast ~> saveTweet
-                                bcast ~> tweetStats ~> Sink.foreach(println)
-
-      ClosedShape
-    })
-    g.run()
-  }
-
   twitterStream()
 }
+
