+++ src/test/scala/tweet/StreamTest.scala
@@ -1,16 +1,16 @@
+package tweet
+
 import akka.actor.ActorSystem
+import akka.stream.scaladsl.Source
+import akka.stream.testkit.scaladsl.TestSink
 import akka.stream.{ActorMaterializer, ActorMaterializerSettings}
-import akka.stream.scaladsl.{Sink, Source}
-import akka.stream.testkit.scaladsl.{TestSink, TestSource}
-import akka.testkit.TestProbe
 import org.json4s.DefaultFormats
 import org.json4s.native.JsonMethods._
 import org.scalatest.{FlatSpec, Matchers}
 
 import scala.util.Failure
 
-
-class TwitterStreamerSpec extends FlatSpec with Matchers {
+class StreamTest extends FlatSpec with Matchers {
   implicit val actorSystem = ActorSystem("test")
   implicit val materializer = ActorMaterializer(ActorMaterializerSettings(actorSystem))
 
@@ -19,9 +19,9 @@ class TwitterStreamerSpec extends FlatSpec with Matchers {
   behavior of "tweetStats"
 
   it should "count correctly" in {
-    val testTweet = TwitterStreamerSpec.randomTweet
+    val testTweet = StreamSpec.randomTweet
     val tweets = List(testTweet.copy(retweeted_status = None), testTweet, testTweet.copy(retweeted_status = None))
-    val testSource = Source(tweets).via(TwitterStreamer.tweetStats)
+    val testSource = Source(tweets).via(Stream.tweetStats)
     testSource
       .runWith(TestSink.probe[String])
       .request(1)
@@ -33,14 +33,14 @@ class TwitterStreamerSpec extends FlatSpec with Matchers {
 
   it should "fail silently" in {
     val invalidTweetData = "gsfsd"
-    TwitterStreamer.jsonExtract(invalidTweetData) shouldBe a [Failure[_]]
+    Stream.jsonExtract(invalidTweetData) shouldBe a [Failure[_]]
   }
 }
 
-object TwitterStreamerSpec {
+object StreamSpec {
   implicit val formats = DefaultFormats
   def randomTweet: Tweet = {
-    val json = scala.io.Source.fromURL(getClass.getResource("tweet.json")).mkString.stripLineEnd
+    val json = scala.io.Source.fromURL(getClass.getResource("/tweet.json")).mkString.stripLineEnd
     parse(json).extract[Tweet]
   }
 }
